fixed type errors
